* 'devel' of https://github.com/arangodb/arangodb: (25 commits)
  Add footer to docs, unify slogan
  Docs: make graph traversal images hard copies (fixes Cygwin builds)
  Added documentation for RocksDB and hash cache options to manual.
  Changed default cache size to something reasonable.
  cppcheck
  added comments
  fix shutdown order
  Added document cache.
  fix some replication issues
  Invalidate current coordinators on every 2nd heartbeatrun
  Show timeout when failing
  Addding locking, fixing conversion warning
  Fixing locking error
  resync changelog to 3.1
  Fix jslint
  CHANGELOG: SIGHUP forwarding
  forward SIG_HUP from supervisor
  Allow removing dbnodes
  Rocksb: index locking issue
  Fix serverAlias map...this was ridiculous
  ...

+++ arangod/RocksDBEngine/RocksDBEdgeIndex.cpp
@@ -69,8 +69,7 @@ RocksDBEdgeIndexIterator::RocksDBEdgeIndexIterator(
       _doUpdateBounds(true),
       _useCache(useCache),
       _cache(cache),
-      _cacheValueSize(0)
-{
+      _cacheValueSize(0) {
   keys.release();  // now we have ownership for _keys
   TRI_ASSERT(_keys->slice().isArray());
   RocksDBTransactionState* state = rocksutils::toRocksTransactionState(_trx);
@@ -130,7 +129,7 @@ bool RocksDBEdgeIndexIterator::next(TokenCallback const& cb, size_t limit) {
       if(_doUpdateBounds){
         LOG_TOPIC(ERR, Logger::FIXME) << "looking for cache key ";
         // try to find cached value
-        auto f = _cache->find(fromTo.data(),fromTo.size());
+        auto f = _cache->find(fromTo.data(), (uint32_t)fromTo.size());
         foundInCache = f.found();
         if (foundInCache) {
           LOG_TOPIC(ERR, Logger::FIXME) << "found in cache " << fromTo;
